I remembered to commit

+++ src/main/java/com/thesis/backend/repository/RoleRepository.java
@@ -1,6 +1,6 @@
 package com.thesis.backend.repository;
 
-import com.thesis.backend.entities.ERole;
+import com.thesis.backend.entities.enums.ERole;
 import com.thesis.backend.entities.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Integer> {
     Optional<Role> findByName(ERole name);
 }
